refactor: extract admin reports tab
